`LineChart` layout fixes (WIP): use max tick-width to calculate left margin

+++ script/build-standalone.sh
@@ -3,11 +3,17 @@
 # set -x
 set -e
 
+NAME=$(basename $1 .purs)
+
+if [[ ! -e "src/Standalone/$NAME.purs" ]]; then
+  echo "Error: '$NAME.purs' not found."
+  exit 1
+fi
+
 # Could factor this through build.sh but currently these are shared by all standalone figures
 cp -r fluid dist/app
 cp -r web/css dist/app
 
-NAME=$(basename $1 .purs)
 NAME_LISP_CASE=$(echo $NAME | sed 's/\([A-Z]\)/-\1/g' | tr 'A-Z' 'a-z' | sed 's/^-//')
 echo "$1 -> $NAME_LISP_CASE"
